% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/open_device.R
\name{open_device}
\alias{open_device}
\title{Open a plotting device}
\usage{
open_device(file, height = 7, width = 7, ppi_png = 300)
}
\arguments{
\item{file}{figure will be written to {file}. Must end in .pdf or .png.}

\item{height}{height of output figure}

\item{width}{width of output figure}

\item{ppi_png}{pixel density (ignored if file ends in .pdf)}
}
\description{
This function opens a PDF or PNG plotting device.
}
