% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/source_partial.R
\name{source_partial}
\alias{source_partial}
\title{Source part of an R script}
\usage{
source_partial(file, start, end, ...)
}
\arguments{
\item{file}{R script to be sourced}

\item{start}{first line to execute}

\item{end}{last line to execute}
}
\description{
This function runs all lines of an R script from a start line through an end line.
https://stackoverflow.com/questions/12214963/source-only-part-of-a-file
}
