% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_sleek.R
\name{theme_sleek}
\alias{theme_sleek}
\title{Sleek theme for ggplot2}
\usage{
theme_sleek(mode = c("dark", "light"))
}
\arguments{
\item{mode}{"dark" or "light"}
}
\description{
This function creates a classic theme with options dark (solarized dark) or light (for papers).
}
