% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_latex_table.R
\name{write_latex_table}
\alias{write_latex_table}
\title{Write LaTex table to file}
\usage{
write_latex_table(table, file)
}
\arguments{
\item{table}{a dataframe to be converted to a L}

\item{file}{a file name (presumably ending in .tex) to which to write the LaTeX table}
}
\description{
This function is necessary because for some reason xtable::print.xtable includes an end-line
character at the end of its tables, which causes a blank row to be appended to the bottom of the
table. The most important feature of this function is that
}
