% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/color.R
\name{color}
\alias{color}
\title{Pick a color}
\usage{
color(
  color = c("axis", "fg", "bg", "base", "base2", "red", "orange", "yellow", "green",
    "blue"),
  mode = c("dark", "light")
)
}
\arguments{
\item{color}{"axis", "fg", "bg", "base", "base2", "red", "orange", "yellow", "green" or "blue"}

\item{mode}{"dark" or "light"}
}
\value{
a character string representing a color
}
\description{
This function picks an appropriate version of a color based on model (dark/light)
}
