% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_latex_table.R
\name{write_latex_table}
\alias{write_latex_table}
\title{Write LaTex table to file}
\usage{
write_latex_table(
  table,
  file,
  buffer_row = NULL,
  hline.after = NULL,
  include.rownames = FALSE,
  include.colnames = FALSE,
  only.contents = TRUE,
  sanitize.text.function = identity,
  ...
)
}
\arguments{
\item{table}{a dataframe to be converted to a L}

\item{file}{a file name (presumably ending in .tex) to which to write the LaTeX table}

\item{buffer_row}{row number before which to insert blank row}

\item{hline.after}{a vector indicating the rows after which a horizontal line should appear.
Default value is NULL.}

\item{include.rownames}{If \code{TRUE} the row names are printed. Default value is \code{FALSE}.}

\item{include.colnames}{If \code{TRUE} the col names are printed. Default value is \code{FALSE}.}

\item{only.contents}{If \verb{TRUE`` only the rows of the table are printed. Default value is }TRUE`.}

\item{...}{additional arugments to \code{print.xtable}}

\item{sanitize.test.function}{A function taking a character vector and returning a sanitized one.
Default value is \code{identity}.}
}
\description{
This function is necessary because for some reason xtable::print.xtable includes an end-line
character at the end of its tables, which causes a blank row to be appended to the bottom of the
table. The most important feature of this function is that
}
